Moves GenerateResponseChunk to its own file.
