+++ js/ai/src/generate.ts
@@ -24,12 +24,11 @@ import {
 import { Registry } from '@genkit-ai/core/registry';
 import { parseSchema, toJsonSchema } from '@genkit-ai/core/schema';
 import { DocumentData } from './document.js';
-import { extractJson } from './extract.js';
+import { GenerateResponseChunk } from './generate/chunk.js';
 import { generateHelper, GenerateUtilParamSchema } from './generateAction.js';
 import { Message } from './message.js';
 import {
   GenerateRequest,
-  GenerateResponseChunkData,
   GenerateResponseData,
   GenerationCommonConfigSchema,
   GenerationUsage,
@@ -207,96 +206,6 @@ export class GenerateResponse<O = unknown> implements ModelResponseData {
     if (!out.finishMessage) delete out.finishMessage;
     if (!out.request) delete out.request;
     return out;
-  }
-}
-
-export class GenerateResponseChunk<T = unknown>
-  implements GenerateResponseChunkData
-{
-  /** The index of the candidate this chunk corresponds to. */
-  index?: number;
-  /** The content generated in this chunk. */
-  content: Part[];
-  /** Custom model-specific data for this chunk. */
-  custom?: unknown;
-  /** Accumulated chunks for partial output extraction. */
-  accumulatedChunks?: GenerateResponseChunkData[];
-
-  constructor(
-    data: GenerateResponseChunkData,
-    accumulatedChunks?: GenerateResponseChunkData[]
-  ) {
-    this.index = data.index;
-    this.content = data.content || [];
-    this.custom = data.custom;
-    this.accumulatedChunks = accumulatedChunks;
-  }
-
-  /**
-   * Concatenates all `text` parts present in the chunk with no delimiter.
-   * @returns A string of all concatenated text parts.
-   */
-  get text(): string {
-    return this.content.map((part) => part.text || '').join('');
-  }
-
-  /**
-   * Concatenates all `text` parts of all chunks from the response thus far.
-   * @returns A string of all concatenated chunk text content.
-   */
-  get accumulatedText(): string {
-    if (!this.accumulatedChunks)
-      throw new GenkitError({
-        status: 'FAILED_PRECONDITION',
-        message: 'Cannot compose accumulated text without accumulated chunks.',
-      });
-
-    return this.accumulatedChunks
-      ?.map((c) => c.content.map((p) => p.text || '').join(''))
-      .join('');
-  }
-
-  /**
-   * Returns the first media part detected in the chunk. Useful for extracting
-   * (for example) an image from a generation expected to create one.
-   * @returns The first detected `media` part in the chunk.
-   */
-  get media(): { url: string; contentType?: string } | null {
-    return this.content.find((part) => part.media)?.media || null;
-  }
-
-  /**
-   * Returns the first detected `data` part of a chunk.
-   * @returns The first `data` part detected in the chunk (if any).
-   */
-  get data(): T | null {
-    return this.content.find((part) => part.data)?.data as T | null;
-  }
-
-  /**
-   * Returns all tool request found in this chunk.
-   * @returns Array of all tool request found in this chunk.
-   */
-  get toolRequests(): ToolRequestPart[] {
-    return this.content.filter(
-      (part) => !!part.toolRequest
-    ) as ToolRequestPart[];
-  }
-
-  /**
-   * Attempts to extract the longest valid JSON substring from the accumulated chunks.
-   * @returns The longest valid JSON substring found in the accumulated chunks.
-   */
-  get output(): T | null {
-    if (!this.accumulatedChunks) return null;
-    const accumulatedText = this.accumulatedChunks
-      .map((chunk) => chunk.content.map((part) => part.text || '').join(''))
-      .join('');
-    return extractJson<T>(accumulatedText, false);
-  }
-
-  toJSON(): GenerateResponseChunkData {
-    return { index: this.index, content: this.content, custom: this.custom };
   }
 }
