+++ js/ai/tests/generate/generate_test.ts
@@ -20,22 +20,22 @@ import { toJsonSchema } from '@genkit-ai/core/schema';
 import assert from 'node:assert';
 import { beforeEach, describe, it } from 'node:test';
 import {
+  generate,
   GenerateOptions,
   GenerateResponse,
-  GenerateResponseChunk,
   GenerationBlockedError,
   GenerationResponseError,
-  generate,
   toGenerateRequest,
 } from '../../src/generate.js';
+import { GenerateResponseChunk } from '../../src/generate/chunk.js';
 import { Message } from '../../src/message.js';
 import {
+  defineModel,
   GenerateRequest,
   GenerateResponseChunkData,
   GenerateResponseData,
   ModelAction,
   ModelMiddleware,
-  defineModel,
 } from '../../src/model.js';
 import { defineTool } from '../../src/tool.js';
